==> analyses/9-compute_ms_metrics.R <==
#'
#' Compute all metrics reported in the manucript
#'

number_of_original_journals()
number_of_original_journals_per_publisher_type()
number_of_original_articles()
number_of_cited_references()

number_of_filtered_journals()
number_of_filtered_journals_per_publisher_type()
number_of_filtered_original_articles()
number_of_filtered_cited_references()

mean_number_cited_references_per_article()


#'
#' Compute all metrics reported in the supplementary material
#'

number_original_dafnee_journals()
number_original_additional_journals()

median_number_of_original_articles_per_journal()
mean_number_of_original_articles_per_journal()

median_number_of_cited_references_per_articles()
mean_number_of_cited_references_per_articles()

number_of_articles_w_refs_in_dafnee_db()
number_of_cited_references_in_dafnee_db()

number_of_filtered_articles_per_publisher_type()
number_of_filtered_references_per_publisher_type()
